requirements.txt
# PaddleCoach - AI Coaching Suite Requirements
# Mohnish's Modules (3A, 3B, 3C)

# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0

# MediaPipe for Pose Estimation (Module 3A)
mediapipe>=0.10.0

# Google Gemini for AI Analysis
google-generativeai>=0.3.0

# Data Processing
dataclasses-json>=0.6.0

# Optional: For better video processing
opencv-contrib-python>=4.8.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Note: Veo 3.1 and Nano Banana APIs don't have official Python packages yet
# These will be accessed via REST APIs using the requests library
